Migrate to biome.js

+++ scripts/screenshotter.mjs
@@ -1,8 +1,8 @@
 #!/usr/bin/env node
 import fs from "fs";
 import path from "path";
-import puppeteer from "puppeteer";
 import url from "url";
+import puppeteer from "puppeteer";
 
 const __dirname = path.dirname(url.fileURLToPath(import.meta.url));
 const pathToExtension = path.resolve(__dirname, "../extension");
@@ -44,19 +44,14 @@ function getFilePath(screenshotsDir, url) {
 }
 
 function prepareDir(screenshotsDir) {
-  if (fs.existsSync(screenshotsDir))
-    fs.rmdirSync(screenshotsDir, { recursive: true });
+  if (fs.existsSync(screenshotsDir)) fs.rmdirSync(screenshotsDir, { recursive: true });
   fs.mkdirSync(screenshotsDir);
 }
 
 async function main(puppeteer) {
   const browser = await puppeteer.launch({
     headless: false,
-    args: [
-      `--disable-extensions-except=${pathToExtension}`,
-      `--load-extension=${pathToExtension}`,
-      "--lang=cs-CZ,cs"
-    ],
+    args: [`--disable-extensions-except=${pathToExtension}`, `--load-extension=${pathToExtension}`, "--lang=cs-CZ,cs"],
     defaultViewport: {
       width: 1920,
       height: 1200
